other/req.py

import os
import tweepy

import discord
from discord.ext import commands
from discord.ext import tasks
from discord.ext.commands import CommandNotFound
from discord.utils import get
from discord import Intents

from layeris.layer_image import LayerImage
from layeris.utils.conversions import convert_float_to_uint
from PIL import Image, ImageDraw

import re
import json
import github3

import requests
from requests import get as requests_get
from urllib.request import urlopen, Request
import cloudscraper

from time import sleep
from io import BytesIO


from bs4 import BeautifulSoup as Soup

import asyncio
import aiosonic
from aiosonic.timeout import Timeouts
import aiohttp
import nest_asyncio
import html5lib


from datetime import datetime, timedelta, date
from dateutil.relativedelta import relativedelta
from pytz import timezone


from threading import Thread
import dictdiffer  
from random import randint
from ast import literal_eval
from pprint import pprint
from dotenv import load_dotenv