other/requirements.txt
html5lib==1.1
pytz==2020.5
aiohttp==3.7.3
dictdiffer==0.8.1
layeris==0.1.3
aiosonic==0.9.5
nest_asyncio==1.5.1
github3.py==1.3.0
requests==2.25.0
discord.py==1.6.0
tweepy==3.10.0
cloudscraper==1.2.56
beautifulsoup4==4.9.3
discord==1.0.1
Pillow==8.1.1
python-dotenv==0.15.0
python_dateutil==2.8.1
